restaubot/menuitem/serializers.py
from rest_framework import serializers
from menuitem.models import MenuItem


class MenuItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = MenuItem
        fields = '__all__'