requirements.txt
psycopg
django
djangorestframework
Pillow
drf-yasg
